requirements.txt
discord
requests
python-dotenv
logbook
requests